// src/front/src/api/rcp/models/monitorEntrySummary.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Rcp API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { RcpTimeEntryStatus } from "./rcpTimeEntryStatus";

export interface MonitorEntrySummary {
	/** @nullable */
	gid?: string | null;
	userId?: number;
	/** @nullable */
	userFullName?: string | null;
	status?: RcpTimeEntryStatus;
	totalHours?: number;
	filledDays?: number;
	/** @nullable */
	lastEntryDate?: string | null;
}